+++ src/advanced-options/advanced-options.component.spec.ts
@@ -1,0 +1,25 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { AdvancedOptionsComponent } from './advanced-options.component';
+
+describe('AdvancedOptionsComponent', () => {
+  let component: AdvancedOptionsComponent;
+  let fixture: ComponentFixture<AdvancedOptionsComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [AdvancedOptionsComponent]
+    })
+      .compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(AdvancedOptionsComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});