+++ CHANGELOG.md
@@ -1,3 +1,42 @@
+## 1.3.3
+
+* Added ecliptic longitude tooltips to the large sun and moon icons, instead of just the smaller true positions markers.
+* Fixed a bug with showing the Advanced Options panel while to Control panel is collapsed.
+
+## 1.3.2
+
+* Additional translations to French and German.
+* Fixed inaccuracies in placement of sun/moon/planet markers.
+* Added tooltips for ecliptic longitude.
+
+## 1.3.1
+
+* Fixed Firefox map drawing bug.
+* Fixed i18n website path formation.
+
+## 1.3.0
+
+* New “Advanced Options” panel.
+* Added animation play/fast-forward/stop buttons.
+* Option added to display greater mechanical detail of the clock.
+* Option added for emulation of mechanical timekeeping (supersedes “Align sun to hand pointer” from version 1.2.0).
+* Option for sun/moon "real position" markers.
+* Option to show addition planet positions (Mercury, Venus, Mars, Jupiter, Saturn).
+* For greater consistency with post-2018 appearance, globe map can be hidden.
+* Collapsed control panel displays currently-set time.
+* Option added for keeping SVG shading at the risk of slower graphics.
+* Added info panel with time readouts, sun/moon rise/set times, and other data.
+
+## 1.2.0
+
+* New color scheme to match October 2018 renovation, with older colors optionally available.
+* Option to align sun icon with hand pointer (more authentic, less accurate sun positioning).
+
+## 1.1.0
+
+* Initial work on internationalization/localization, with Czech and Spanish translations.
+* Control panel is now collapsible.
+
 ## 1.0.1
 
 * Updated color theme for controls.